Python: Fixed vh.classmc, pipes; changed vh.CloseConnection
Forgot that vh.classmc had a third argument added (optional) and broke it, but now it is fixed. Added checks so that no pipe is appended if the message already ends with one. Added optional third argument (close reason) to vh.CloseConnection, so now the default reason for vh.CloseConnection is eCR_DEFAULT and not eCR_KICKED.